Vector index deletion leverages the LSM-tree anti-matter pattern, where delete operations insert tombstone tuples that reconcile with existing records during merge operations.
Mechanism:
  - Anti-matter tuple format: Same as data tuple <distance_to_centroid, cos(θ), vector, primary_key> with anti-matter bit set
  - Reconciliation during merge: When merging LSM components, tuples with matching primary keys annihilate each other when one has the anti-matter bit set
  - Cluster-aware deletion: Anti-matter tuples are inserted into the same cluster as the original vector (using vector value to navigate static structure)

  Add support for inserting records into LSMVCTree memory components and
  flushing to disk components. This enables dynamic record insertion into
  the hierarchical IVF vector index structure.

  Key changes:

  * LSMVCTree insertion path:
    - Insert records into memory component with automatic static structure
      initialization on first insert
    - Copy static structure (centroids, tree navigation pages) from static
      structure component to memory component via setStaticStructure()
    - Delegate to VectorClusteringTree for actual insertion operations

  * VectorClusteringTree insertion logic:
    - Find closest leaf centroid using tree navigation
    - Create/update metadata pages for cluster distance ranges
    - Insert data tuples with format: <distance:ADOUBLE, field:AINTEGER, pk:ALONG>
    - Handle page chaining for clusters spanning multiple data pages

  * Memory to disk flush (doFlush):
    - Copy all pages (including page 0 metadata) from memory component
    - Create self-contained disk components via LSMVCTreeFlushLoader
    - Preserve static structure and data pages in flushed component

  * Multi-cluster search enhancements:
    - VectorClusteringSearchCursor probes multiple clusters per component
      until K records collected
    - LSMVCTreeSearchCursor coordinates sequential search across multiple
      LSM components (memory + disk)
    - DFS-based tree traversal finds next closest clusters in distance order
    - Skip empty clusters during component iteration

+++ hyracks-fullstack/hyracks/hyracks-storage-am-btree/src/main/java/org/apache/hyracks/storage/am/vector/api/IVectorClusteringMetadataFrame.java
@@ -54,7 +54,7 @@ public interface IVectorClusteringMetadataFrame extends IVectorClusteringFrame {
      * @return the maximum distance
      * @throws HyracksDataException if an error occurs
      */
-    float getMaxDistance(int tupleIndex) throws HyracksDataException;
+    double getMaxDistance(int tupleIndex) throws HyracksDataException;
 
     /**
      * Gets the data page pointer for a metadata entry.
